+++ Sources/Yams/Emitter.swift
@@ -7,7 +7,7 @@
 //
 
 #if SWIFT_PACKAGE
-import CYaml
+@_implementationOnly import CYaml
 #endif
 import Foundation
 
@@ -15,7 +15,7 @@ import Foundation
 ///
 /// - parameter objects:       Sequence of Objects.
 /// - parameter canonical:     Output should be the "canonical" format as in the YAML specification.
-/// - parameter indent:        The intendation increment.
+/// - parameter indent:        The indentation increment.
 /// - parameter width:         The preferred line width. @c -1 means unlimited.
 /// - parameter allowUnicode:  Unescaped non-ASCII characters are allowed if true.
 /// - parameter lineBreak:     Preferred line break.
@@ -23,6 +23,8 @@ import Foundation
 /// - parameter explicitEnd:   Explicit document end `...`.
 /// - parameter version:       YAML version directive.
 /// - parameter sortKeys:      Whether or not to sort Mapping keys in lexicographic order.
+/// - parameter sequenceStyle: The style for sequences (arrays / lists)
+/// - parameter mappingStyle:  The style for mappings (dictionaries)
 ///
 /// - returns: YAML string.
 ///
@@ -37,7 +39,9 @@ public func dump<Objects>(
     explicitStart: Bool = false,
     explicitEnd: Bool = false,
     version: (major: Int, minor: Int)? = nil,
-    sortKeys: Bool = false) throws -> String
+    sortKeys: Bool = false,
+    sequenceStyle: Node.Sequence.Style = .any,
+    mappingStyle: Node.Mapping.Style = .any) throws -> String
     where Objects: Sequence {
     func representable(from object: Any) throws -> NodeRepresentable {
         if let representable = object as? NodeRepresentable {
@@ -56,7 +60,9 @@ public func dump<Objects>(
         explicitStart: explicitStart,
         explicitEnd: explicitEnd,
         version: version,
-        sortKeys: sortKeys
+        sortKeys: sortKeys,
+        sequenceStyle: sequenceStyle,
+        mappingStyle: mappingStyle
     )
 }
 
@@ -64,7 +70,7 @@ public func dump<Objects>(
 ///
 /// - parameter object:        Object.
 /// - parameter canonical:     Output should be the "canonical" format as in the YAML specification.
-/// - parameter indent:        The intendation increment.
+/// - parameter indent:        The indentation increment.
 /// - parameter width:         The preferred line width. @c -1 means unlimited.
 /// - parameter allowUnicode:  Unescaped non-ASCII characters are allowed if true.
 /// - parameter lineBreak:     Preferred line break.
@@ -72,6 +78,8 @@ public func dump<Objects>(
 /// - parameter explicitEnd:   Explicit document end `...`.
 /// - parameter version:       YAML version directive.
 /// - parameter sortKeys:      Whether or not to sort Mapping keys in lexicographic order.
+/// - parameter sequenceStyle: The style for sequences (arrays / lists)
+/// - parameter mappingStyle:  The style for mappings (dictionaries)
 ///
 /// - returns: YAML string.
 ///
@@ -86,7 +94,9 @@ public func dump(
     explicitStart: Bool = false,
     explicitEnd: Bool = false,
     version: (major: Int, minor: Int)? = nil,
-    sortKeys: Bool = false) throws -> String {
+    sortKeys: Bool = false,
+    sequenceStyle: Node.Sequence.Style = .any,
+    mappingStyle: Node.Mapping.Style = .any) throws -> String {
     return try serialize(
         node: object.represented(),
         canonical: canonical,
@@ -97,7 +107,9 @@ public func dump(
         explicitStart: explicitStart,
         explicitEnd: explicitEnd,
         version: version,
-        sortKeys: sortKeys
+        sortKeys: sortKeys,
+        sequenceStyle: sequenceStyle,
+        mappingStyle: mappingStyle
     )
 }
 
@@ -105,7 +117,7 @@ public func dump(
 ///
 /// - parameter nodes:         Sequence of `Node`s.
 /// - parameter canonical:     Output should be the "canonical" format as in the YAML specification.
-/// - parameter indent:        The intendation increment.
+/// - parameter indent:        The indentation increment.
 /// - parameter width:         The preferred line width. @c -1 means unlimited.
 /// - parameter allowUnicode:  Unescaped non-ASCII characters are allowed if true.
 /// - parameter lineBreak:     Preferred line break.
@@ -113,6 +125,8 @@ public func dump(
 /// - parameter explicitEnd:   Explicit document end `...`.
 /// - parameter version:       YAML version directive.
 /// - parameter sortKeys:      Whether or not to sort Mapping keys in lexicographic order.
+/// - parameter sequenceStyle: The style for sequences (arrays / lists)
+/// - parameter mappingStyle:  The style for mappings (dictionaries)
 ///
 /// - returns: YAML string.
 ///
@@ -127,7 +141,9 @@ public func serialize<Nodes>(
     explicitStart: Bool = false,
     explicitEnd: Bool = false,
     version: (major: Int, minor: Int)? = nil,
-    sortKeys: Bool = false) throws -> String
+    sortKeys: Bool = false,
+    sequenceStyle: Node.Sequence.Style = .any,
+    mappingStyle: Node.Mapping.Style = .any) throws -> String
     where Nodes: Sequence, Nodes.Iterator.Element == Node {
     let emitter = Emitter(
         canonical: canonical,
@@ -138,7 +154,9 @@ public func serialize<Nodes>(
         explicitStart: explicitStart,
         explicitEnd: explicitEnd,
         version: version,
-        sortKeys: sortKeys
+        sortKeys: sortKeys,
+        sequenceStyle: sequenceStyle,
+        mappingStyle: mappingStyle
     )
     try emitter.open()
     try nodes.forEach(emitter.serialize)
@@ -150,7 +168,7 @@ public func serialize<Nodes>(
 ///
 /// - parameter node:          `Node`.
 /// - parameter canonical:     Output should be the "canonical" format as in the YAML specification.
-/// - parameter indent:        The intendation increment.
+/// - parameter indent:        The indentation increment.
 /// - parameter width:         The preferred line width. @c -1 means unlimited.
 /// - parameter allowUnicode:  Unescaped non-ASCII characters are allowed if true.
 /// - parameter lineBreak:     Preferred line break.
@@ -158,6 +176,8 @@ public func serialize<Nodes>(
 /// - parameter explicitEnd:   Explicit document end `...`.
 /// - parameter version:       YAML version directive.
 /// - parameter sortKeys:      Whether or not to sort Mapping keys in lexicographic order.
+/// - parameter sequenceStyle: The style for sequences (arrays / lists)
+/// - parameter mappingStyle:  The style for mappings (dictionaries)
 ///
 /// - returns: YAML string.
 ///
@@ -172,7 +192,9 @@ public func serialize(
     explicitStart: Bool = false,
     explicitEnd: Bool = false,
     version: (major: Int, minor: Int)? = nil,
-    sortKeys: Bool = false) throws -> String {
+    sortKeys: Bool = false,
+    sequenceStyle: Node.Sequence.Style = .any,
+    mappingStyle: Node.Mapping.Style = .any) throws -> String {
     return try serialize(
         nodes: [node],
         canonical: canonical,
@@ -183,7 +205,9 @@ public func serialize(
         explicitStart: explicitStart,
         explicitEnd: explicitEnd,
         version: version,
-        sortKeys: sortKeys
+        sortKeys: sortKeys,
+        sequenceStyle: sequenceStyle,
+        mappingStyle: mappingStyle
     )
 }
 
@@ -224,6 +248,12 @@ public final class Emitter {
 
         /// Set if emitter should sort keys in lexicographic order.
         public var sortKeys: Bool = false
+
+        /// Set the style for sequences (arrays / lists)
+        public var sequenceStyle: Node.Sequence.Style = .any
+
+        /// Set the style for mappings (dictionaries)
+        public var mappingStyle: Node.Mapping.Style = .any
     }
 
     /// Configuration options to use when emitting YAML.
@@ -245,6 +275,8 @@ public final class Emitter {
     /// - parameter explicitEnd:   Explicit document end `...`.
     /// - parameter version:       The `%YAML` directive value or nil.
     /// - parameter sortKeys:      Set if emitter should sort keys in lexicographic order.
+    /// - parameter sequenceStyle: Set the style for sequences (arrays / lists)
+    /// - parameter mappingStyle:  Set the style for mappings (dictionaries)
     public init(canonical: Bool = false,
                 indent: Int = 0,
                 width: Int = 0,
@@ -253,7 +285,9 @@ public final class Emitter {
                 explicitStart: Bool = false,
                 explicitEnd: Bool = false,
                 version: (major: Int, minor: Int)? = nil,
-                sortKeys: Bool = false) {
+                sortKeys: Bool = false,
+                sequenceStyle: Node.Sequence.Style = .any,
+                mappingStyle: Node.Mapping.Style = .any) {
         options = Options(canonical: canonical,
                           indent: indent,
                           width: width,
@@ -262,7 +296,9 @@ public final class Emitter {
                           explicitStart: explicitStart,
                           explicitEnd: explicitEnd,
                           version: version,
-                          sortKeys: sortKeys)
+                          sortKeys: sortKeys,
+                          sequenceStyle: sequenceStyle,
+                          mappingStyle: mappingStyle)
         // configure emitter
         yaml_emitter_initialize(&emitter)
         yaml_emitter_set_output(&self.emitter, { pointer, buffer, size in
@@ -281,7 +317,7 @@ public final class Emitter {
         yaml_emitter_delete(&emitter)
     }
 
-    /// Open & initialize the emmitter.
+    /// Open & initialize the emitter.
     ///
     /// - throws: `YamlError` if the `Emitter` was already opened or closed.
     public func open() throws {
@@ -379,9 +415,12 @@ extension Emitter.Options {
     /// - parameter explicitEnd:   Explicit document end `...`.
     /// - parameter version:       The `%YAML` directive value or nil.
     /// - parameter sortKeys:      Set if emitter should sort keys in lexicographic order.
+    /// - parameter sequenceStyle: Set the style for sequences (arrays / lists)
+    /// - parameter mappingStyle:  Set the style for mappings (dictionaries)
     public init(canonical: Bool = false, indent: Int = 0, width: Int = 0, allowUnicode: Bool = false,
                 lineBreak: Emitter.LineBreak = .ln, version: (major: Int, minor: Int)? = nil,
-                sortKeys: Bool = false) {
+                sortKeys: Bool = false, sequenceStyle: Node.Sequence.Style = .any,
+                mappingStyle: Node.Mapping.Style = .any) {
         self.canonical = canonical
         self.indent = indent
         self.width = width
@@ -389,6 +428,8 @@ extension Emitter.Options {
         self.lineBreak = lineBreak
         self.version = version
         self.sortKeys = sortKeys
+        self.sequenceStyle = sequenceStyle
+        self.mappingStyle = mappingStyle
     }
 }